Sprint3Final

+++ src/view/Window.java
@@ -1,3 +1,8 @@
+/*
+ * Final Project
+ *
+ * Winter 2023
+ */
 package view;
 
 import java.awt.*;
@@ -6,13 +11,16 @@ import java.awt.event.ActionListener;
 import java.awt.event.KeyAdapter;
 import java.awt.event.KeyEvent;
 import javax.swing.*;
-import model.BoardControls;
+import controls.BoardControls;
 
 /**
- * This object represents the frame of the GUI.
+ * This object represents the upper-most frame of the GUI and lists
+ * out its contents.
  *
  * @author Avreen Kaur
- * @author Faith
+ * @author Faith Capito
+ * @author Sullivan Seamus Lucier-Benson
+ * @author Josh Chang
  *
  * @version Winter 2023
  */
@@ -41,11 +49,6 @@ public class Window extends JFrame
     private static final int TIMER_DELAY = 1000;
 
     /**
-     * Timer that controls how game functions.
-     */
-    protected final Timer myTimer;
-
-    /**
      * This is a red panel object.
      */
     private RedPanel myRed = new RedPanel();
@@ -68,7 +71,7 @@ public class Window extends JFrame
     /**
      * File menu at top of window.
      */
-    private FileMenu myFileMenu = new FileMenu();
+    private FileMenu myFileMenu;
 
     /**
      * This represents a Board Object of the model package.
@@ -76,16 +79,19 @@ public class Window extends JFrame
     private final BoardControls myBoard;
 
     /**
-     * This is the drawBoardPanel object which represents the board
-     * of the game as a drawing in Red Panel.
+     * Keeps track of whether the game has been paused or not.
      */
-    private final DrawBoardPanel myBoardPanel = new DrawBoardPanel();
+    private boolean myGameInProgress;
 
     /**
-     * This is the drawPiecePanel object which represents the next piece
-     * of the game as a drawing in Green Panel.
+     * Keeps track of whether the game can be started again or not.
      */
-    private final DrawPiecePanel myPiecePanel = new DrawPiecePanel();
+    private boolean myPressToStart;
+
+    /**
+     * Timer that controls how game functions.
+     */
+    protected static Timer myTimer;
 
     // CONSTRUCTORS
 
@@ -99,21 +105,22 @@ public class Window extends JFrame
         myBoard = theBoard;
         myBoard.newGame();
 
+        myFileMenu = new FileMenu();
+
+        myGameInProgress = false;
+        myPressToStart = true;
+
         myWindow = new JFrame(NAME);
         myWindow.setLayout(new BorderLayout());
 
-        myRed.add(myBoardPanel);
-        myGreen.add(myPiecePanel);
-
-        myBlue.getPanel().add(myGreen.getPanel(), BorderLayout.NORTH);
-        myRed.getPanel().add(myBlue.getPanel(), BorderLayout.EAST);
-
         myWindow.add(myFileMenu, BorderLayout.NORTH);
-        myWindow.add(myRed.getPanel(), BorderLayout.CENTER);
+        myBlue.add(myGreen, BorderLayout.NORTH);
+        myRed.add(myBlue, BorderLayout.EAST);
+        myWindow.add(myRed, BorderLayout.CENTER);
 
         myWindow.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
         myWindow.setSize(new Dimension(WIDTH_DIM, LENGTH_DIM));
-        // myWindow.setResizable(false);
+        myWindow.setResizable(false);
         myWindow.setVisible(true);
 
         myTimer = new Timer(TIMER_DELAY, null);
@@ -122,12 +129,17 @@ public class Window extends JFrame
             @Override
             public void actionPerformed(final ActionEvent theEve)
             {
-                myBoard.step();
+                if (myTimer.isRunning())
+                {
+                    myBoard.step();
+                }
             }
         });
 
-        myBoard.addPropertyChangeListener(myBoardPanel);
-        myBoard.addPropertyChangeListener(myPiecePanel);
+        myBoard.addPropertyChangeListener(myRed);
+        myBoard.addPropertyChangeListener(myGreen);
+        myBoard.addPropertyChangeListener(myBlue);
+        myBoard.addPropertyChangeListener(myFileMenu);
 
         myWindow.addKeyListener(new BoardKeyListener());
         myWindow.setFocusable(true);
@@ -140,33 +152,71 @@ public class Window extends JFrame
         @Override
         public void keyPressed(final KeyEvent theEvent)
         {
-            if (theEvent.getKeyCode() == KeyEvent.VK_A
-                    || theEvent.getKeyCode() == KeyEvent.VK_LEFT)
+
+            if (myPressToStart)
             {
-                myBoard.left();
+                if (theEvent.getKeyCode() == KeyEvent.VK_1) // start the game
+                {
+                    myTimer.start();
+                    myGameInProgress = true; // game has started, control the pieces
+                    myPressToStart = false;
+                }
             }
 
-            else if (theEvent.getKeyCode() == KeyEvent.VK_D
-                    || theEvent.getKeyCode() == KeyEvent.VK_RIGHT)
+            else if (theEvent.getKeyCode() == KeyEvent.VK_P) // pause the game
             {
-                myBoard.right();
+                if (myTimer.isRunning())
+                {
+                    myTimer.stop();
+                    myGameInProgress = false; // no controlling the piece
+                }
+            }
+            else if (theEvent.getKeyCode() == KeyEvent.VK_U) // unpause the game
+            {
+                if (!myTimer.isRunning())
+                {
+                    myTimer.start();
+                    myGameInProgress = true; // can control the piece
+                }
             }
 
-            else if (theEvent.getKeyCode() == KeyEvent.VK_W
-                    || theEvent.getKeyCode() == KeyEvent.VK_UP)
+            if (myGameInProgress)
             {
-                myBoard.rotateCW();
-            }
+                if (theEvent.getKeyCode() == KeyEvent.VK_A
+                        || theEvent.getKeyCode() == KeyEvent.VK_LEFT)
+                {
+                    myBoard.left();
+                }
+                else if (theEvent.getKeyCode() == KeyEvent.VK_D
+                        || theEvent.getKeyCode() == KeyEvent.VK_RIGHT)
+                {
+                    myBoard.right();
+                }
+                else if (theEvent.getKeyCode() == KeyEvent.VK_W
+                        || theEvent.getKeyCode() == KeyEvent.VK_UP)
+                {
+                    myBoard.rotateCW();
+                }
+                else if (theEvent.getKeyCode() == KeyEvent.VK_E)
+                {
+                    myBoard.rotateCCW();
+                }
+                else if (theEvent.getKeyCode() == KeyEvent.VK_S
+                        || theEvent.getKeyCode() == KeyEvent.VK_DOWN)
+                {
+                    myBoard.down();
+                }
+                else if (theEvent.getKeyCode() == KeyEvent.VK_SPACE)
+                {
+                    myBoard.drop();
+                }
 
-            else if (theEvent.getKeyCode() == KeyEvent.VK_S
-                    || theEvent.getKeyCode() == KeyEvent.VK_DOWN)
-            {
-                myBoard.down();
-            }
-
-            else if (theEvent.getKeyCode() == KeyEvent.VK_SPACE)
-            {
-                myBoard.drop();
+                else if (theEvent.getKeyCode() == KeyEvent.VK_2)
+                {
+                    myBoard.setGameOver();
+                    myGameInProgress = false;
+                    myTimer.stop();
+                }
             }
         }
     }
